Time: 0 ms (100%), Space: 12.3 MB (44.93%) - LeetHub

+++ 0797-rabbits-in-forest/0797-rabbits-in-forest.cpp
@@ -5,14 +5,14 @@ public:
 
         //Count freq of ans
         //Lets store all ans in an map
-        unordered_map<float,float> ansFreq;
+        unordered_map<int,int> ansFreq;
         int minCount=0, multiplier;
 
         for(int i=0; i<answers.size(); i++) ansFreq[answers[i]]++;
 
         for(auto pair: ansFreq){
             multiplier=0;
-            multiplier= ceil((pair.second)/(pair.first+1));
+            multiplier= ceil((float)(pair.second)/(pair.first+1));
 
             minCount+=multiplier*(pair.first+1);
         }
